Create dynamic AzureAD distributiongroups
The script allows to create a number of dynamic distributiongroups (up to 256) in AzureAD, where the users are evenly distributed on a best effort basis (based on their random object id).
Active AzureAD groups can also be queried and deleted.
The groups can be used e.g. in large enterprises for Veeam Microsoft365 Backup.
For better description see the script header.